pyproject.toml
[project]
name = "mcp-auditor"
version = "0.1.0"
description = "Agentic security testing for MCP servers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "Roman Mkrtchian", email = "mkrtchia@gmail.com" }]
keywords = ["mcp", "security", "fuzzing", "audit", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "langgraph>=1.0",
    "langchain-anthropic>=1.0",
    "langchain-google-genai>=4.1",
    "langchain-core>=1.0",
    "mcp>=1.0.0",
    "pydantic>=2.0",
    "langgraph-checkpoint-sqlite>=3.0",
    "click>=8.1",
    "rich>=13.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/mkrtchian/mcp-auditor"
Repository = "https://github.com/mkrtchian/mcp-auditor"
Issues = "https://github.com/mkrtchian/mcp-auditor/issues"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.0",
    "ruff>=0.15",
    "pyright>=1.1",
    "langsmith>=0.3",
]

[project.scripts]
mcp-auditor = "mcp_auditor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_auditor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
    "ignore:Core Pydantic V1 functionality:UserWarning",
    "ignore:'_UnionGenericAlias' is deprecated:DeprecationWarning:google.genai",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/mcp_auditor/graph/prompts.py" = ["E501"]
"src/mcp_auditor/domain/category_guidance.py" = ["E501"]
"src/mcp_auditor/cli.py" = ["E402"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
